refactor: 노드 관련 로직 LinkedList 클래스로 이동
